// src/app/shared/models/event-details-response.model.ts
export interface EventDetailsResponse {
    id: number;
    name: string;
    capacity: number;
    description: string;
    createdAt: string;
    updatedAt: string;
    imagePath: string;
    categoryName: string;
    locationName: string;
    cityName: string;
    countryName: String;
    priceValue: number;
}